refactor moviedetails component
