+++ src/components/styles/relatedMovies.css
@@ -33,11 +33,13 @@ img {
   box-shadow: 1px 1px 6px rgba(148, 148, 148, 0.973) inset;
 }
 .related__item > div {
+  width: 100%;
   display: flex;
   flex-direction: column;
-  justify-content: flex-start;
+  justify-content: center;
   align-items: center;
 }
+
 @media screen and (max-width: 600px) {
   .related__item {
     flex-direction: column;
